LPS-89604 Rename to SiteInitializerItem, since it is not a display context and move to util class
