+++ modules/apps/site/site-admin-web/src/main/java/com/liferay/site/admin/web/internal/display/context/comparator/SiteInitializerNameComparator.java
@@ -15,13 +15,13 @@
 package com.liferay.site.admin.web.internal.display.context.comparator;
 
 import com.liferay.portal.kernel.util.OrderByComparator;
-import com.liferay.site.admin.web.internal.display.context.SiteInitializerItemDisplayContext;
+import com.liferay.site.admin.web.internal.util.SiteInitializerItem;
 
 /**
  * @author Marco Leo
  */
 public class SiteInitializerNameComparator
-	extends OrderByComparator<SiteInitializerItemDisplayContext> {
+	extends OrderByComparator<SiteInitializerItem> {
 
 	public SiteInitializerNameComparator() {
 		this(false);
@@ -33,8 +33,8 @@ public class SiteInitializerNameComparator
 
 	@Override
 	public int compare(
-		SiteInitializerItemDisplayContext siteInitializerItemDisplayContext1,
-		SiteInitializerItemDisplayContext siteInitializerItemDisplayContext2) {
+		SiteInitializerItem siteInitializerItemDisplayContext1,
+		SiteInitializerItem siteInitializerItemDisplayContext2) {
 
 		String name1 = siteInitializerItemDisplayContext1.getName();
 		String name2 = siteInitializerItemDisplayContext2.getName();
