+++ modules/apps/site/site-admin-web/src/main/resources/META-INF/resources/init.jsp
@@ -113,8 +113,8 @@ page import="com.liferay.site.admin.web.internal.constants.SiteAdminPortletKeys"
 page import="com.liferay.site.admin.web.internal.constants.SiteAdminWebKeys" %><%@
 page import="com.liferay.site.admin.web.internal.display.context.SiteAdminDisplayContext" %><%@
 page import="com.liferay.site.admin.web.internal.display.context.SiteAdminManagementToolbarDisplayContext" %><%@
-page import="com.liferay.site.admin.web.internal.display.context.SiteInitializerItemDisplayContext" %><%@
 page import="com.liferay.site.admin.web.internal.servlet.taglib.clay.SiteVerticalCard" %><%@
+page import="com.liferay.site.admin.web.internal.util.SiteInitializerItem" %><%@
 page import="com.liferay.sites.kernel.util.Sites" %><%@
 page import="com.liferay.sites.kernel.util.SitesUtil" %>
